Remove circular dependency to NessieClient in Spark SQL extensions and GC modules
The problem is that the Spark SQL extentions for example rely on iceberg-spark3-runtime, which in turn is compiled against a Nessie version that isn't necessarily the same that we use here. Now when the tests run, we run against a nessie-client version that is different from Iceberg and stuff breaks.

The circular dependecy is fixed by introducing the same NessieClient
version that Iceberg uses.

This also moves NessieClient instantiation into subclasses of
AbstractSparkTest so that each subclass can control how to instantiate
the NessieClient (depending on the Nessie version they pull in)

+++ clients/spark-extensions/src/test/java/org/projectnessie/spark/extensions/ITNessieStatements.java
@@ -34,6 +34,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.projectnessie.client.NessieClient;
 import org.projectnessie.client.tests.AbstractSparkTest;
 import org.projectnessie.error.BaseNessieClientServerException;
 import org.projectnessie.error.NessieConflictException;
@@ -52,6 +53,7 @@ public class ITNessieStatements extends AbstractSparkTest {
 
   private String hash;
   private final String refName = "testBranch";
+  private NessieClient nessieClient;
 
   @BeforeAll
   protected static void createDelta() {
@@ -60,22 +62,29 @@ public class ITNessieStatements extends AbstractSparkTest {
   }
 
   @BeforeEach
-  void getHash() throws NessieNotFoundException {
+  void createClientAndGetHash() throws NessieNotFoundException {
+    nessieClient = NessieClient.builder().withUri(url).build();
     hash = nessieClient.getTreeApi().getDefaultBranch().getHash();
   }
 
   @AfterEach
   void removeBranches() throws NessieConflictException, NessieNotFoundException {
-    for (String s : Arrays.asList(refName, "main")) {
-      try {
-        nessieClient
-            .getTreeApi()
-            .deleteBranch(s, nessieClient.getTreeApi().getReferenceByName(s).getHash());
-      } catch (BaseNessieClientServerException e) {
-        // pass
+    if (null != nessieClient) {
+
+      for (String s : Arrays.asList(refName, "main")) {
+        try {
+          nessieClient
+              .getTreeApi()
+              .deleteBranch(s, nessieClient.getTreeApi().getReferenceByName(s).getHash());
+        } catch (BaseNessieClientServerException e) {
+          // pass
+        }
       }
+      nessieClient.getTreeApi().createReference(Branch.of("main", null));
+
+      nessieClient.close();
+      nessieClient = null;
     }
-    nessieClient.getTreeApi().createReference(Branch.of("main", null));
   }
 
   @Test
